refactor: Move view and Rest logic to different controllers
