+++ src/main/kotlin/ecommerce/controller/ProductController.kt
@@ -3,8 +3,6 @@ package ecommerce.controller
 import ecommerce.model.Product
 import ecommerce.repository.ProductRepository
 import org.springframework.http.ResponseEntity
-import org.springframework.stereotype.Controller
-import org.springframework.ui.Model
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
@@ -12,38 +10,30 @@ import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.PutMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
 import java.net.URI
 
-@Controller
+@RestController
 @RequestMapping("/products")
 class ProductController {
+    @GetMapping
+    fun getProducts(): ResponseEntity<List<Product>> {
+        return ResponseEntity.ok().body(ProductRepository().findAll())
+    }
+
+    @GetMapping("/{id}")
+    fun getProductById(
+        @PathVariable("id") id: Long,
+    ): ResponseEntity<Product> {
+        return ResponseEntity.ok().body(ProductRepository().findById(id))
+    }
+
     @PostMapping("")
     fun create(
         @RequestBody product: Product,
     ): ResponseEntity<Void> {
         ProductRepository().save(product)
         return ResponseEntity.created(URI.create("/products/")).build()
-    }
-
-    @GetMapping("/new")
-    fun showCreateForm(): String {
-        return "create_product_form"
-    }
-
-    @GetMapping("")
-    fun read(model: Model): String {
-        model.addAttribute("products", ProductRepository().findAll())
-        return "products"
-    }
-
-    @GetMapping("/edit/{id}")
-    fun showUpdateForm(
-        @PathVariable("id") id: Long,
-        model: Model,
-    ): String {
-        val product = ProductRepository().findById(id)
-        model.addAttribute("product", product)
-        return "edit_product_form"
     }
 
     @PutMapping("/{id}")
